+++ common/src/main/java/com/alien/common/registry/init/item/AlienChitinBlockItems.java
@@ -23,6 +23,11 @@ public class AlienChitinBlockItems {
         AlienChitinBlocks.ABERRANT_CHITIN_BLOCK_STAIRS
     );
 
+    public static final AVPDeferredHolder<BlockItem> ABERRANT_CHITIN_BLOCK_WALL = AVPBlockItems.register(
+        "aberrant_chitin_block_wall",
+        AlienChitinBlocks.ABERRANT_CHITIN_BLOCK_WALL
+    );
+
     public static final AVPDeferredHolder<BlockItem> ABERRANT_CHITIN_BRICKS = AVPBlockItems.register(
         "aberrant_chitin_bricks",
         AlienChitinBlocks.ABERRANT_CHITIN_BRICKS
@@ -36,6 +41,11 @@ public class AlienChitinBlockItems {
     public static final AVPDeferredHolder<BlockItem> ABERRANT_CHITIN_BRICK_STAIRS = AVPBlockItems.register(
         "aberrant_chitin_brick_stairs",
         AlienChitinBlocks.ABERRANT_CHITIN_BRICK_STAIRS
+    );
+
+    public static final AVPDeferredHolder<BlockItem> ABERRANT_CHITIN_BRICK_WALL = AVPBlockItems.register(
+        "aberrant_chitin_brick_wall",
+        AlienChitinBlocks.ABERRANT_CHITIN_BRICK_WALL
     );
 
     public static final AVPDeferredHolder<BlockItem> CHISELED_ABERRANT_CHITIN_BRICKS = AVPBlockItems.register(
@@ -83,6 +93,11 @@ public class AlienChitinBlockItems {
         AlienChitinBlocks.CHITIN_BLOCK_STAIRS
     );
 
+    public static final AVPDeferredHolder<BlockItem> CHITIN_BLOCK_WALL = AVPBlockItems.register(
+        "chitin_block_wall",
+        AlienChitinBlocks.CHITIN_BLOCK_WALL
+    );
+
     public static final AVPDeferredHolder<BlockItem> CHITIN_BRICKS = AVPBlockItems.register(
         "chitin_bricks",
         AlienChitinBlocks.CHITIN_BRICKS
@@ -96,6 +111,11 @@ public class AlienChitinBlockItems {
     public static final AVPDeferredHolder<BlockItem> CHITIN_BRICK_STAIRS = AVPBlockItems.register(
         "chitin_brick_stairs",
         AlienChitinBlocks.CHITIN_BRICK_STAIRS
+    );
+
+    public static final AVPDeferredHolder<BlockItem> CHITIN_BRICK_WALL = AVPBlockItems.register(
+        "chitin_brick_wall",
+        AlienChitinBlocks.CHITIN_BRICK_WALL
     );
 
     public static final AVPDeferredHolder<BlockItem> NETHER_CHITIN_BLOCK = AVPBlockItems.register(
@@ -113,6 +133,11 @@ public class AlienChitinBlockItems {
         AlienChitinBlocks.NETHER_CHITIN_BLOCK_STAIRS
     );
 
+    public static final AVPDeferredHolder<BlockItem> NETHER_CHITIN_BLOCK_WALL = AVPBlockItems.register(
+        "nether_chitin_block_wall",
+        AlienChitinBlocks.NETHER_CHITIN_BLOCK_WALL
+    );
+
     public static final AVPDeferredHolder<BlockItem> NETHER_CHITIN_BRICKS = AVPBlockItems.register(
         "nether_chitin_bricks",
         AlienChitinBlocks.NETHER_CHITIN_BRICKS
@@ -126,6 +151,11 @@ public class AlienChitinBlockItems {
     public static final AVPDeferredHolder<BlockItem> NETHER_CHITIN_BRICK_STAIRS = AVPBlockItems.register(
         "nether_chitin_brick_stairs",
         AlienChitinBlocks.NETHER_CHITIN_BRICK_STAIRS
+    );
+
+    public static final AVPDeferredHolder<BlockItem> NETHER_CHITIN_BRICK_WALL = AVPBlockItems.register(
+        "nether_chitin_brick_wall",
+        AlienChitinBlocks.NETHER_CHITIN_BRICK_WALL
     );
 
     public static final AVPDeferredHolder<BlockItem> POLISHED_ABERRANT_CHITIN = AVPBlockItems.register(
@@ -143,6 +173,11 @@ public class AlienChitinBlockItems {
         AlienChitinBlocks.POLISHED_ABERRANT_CHITIN_STAIRS
     );
 
+    public static final AVPDeferredHolder<BlockItem> POLISHED_ABERRANT_CHITIN_WALL = AVPBlockItems.register(
+        "polished_aberrant_chitin_wall",
+        AlienChitinBlocks.POLISHED_ABERRANT_CHITIN_WALL
+    );
+
     public static final AVPDeferredHolder<BlockItem> POLISHED_CHITIN = AVPBlockItems.register(
         "polished_chitin",
         AlienChitinBlocks.POLISHED_CHITIN
@@ -158,6 +193,11 @@ public class AlienChitinBlockItems {
         AlienChitinBlocks.POLISHED_CHITIN_STAIRS
     );
 
+    public static final AVPDeferredHolder<BlockItem> POLISHED_CHITIN_WALL = AVPBlockItems.register(
+        "polished_chitin_wall",
+        AlienChitinBlocks.POLISHED_CHITIN_WALL
+    );
+
     public static final AVPDeferredHolder<BlockItem> POLISHED_NETHER_CHITIN = AVPBlockItems.register(
         "polished_nether_chitin",
         AlienChitinBlocks.POLISHED_NETHER_CHITIN
@@ -171,6 +211,11 @@ public class AlienChitinBlockItems {
     public static final AVPDeferredHolder<BlockItem> POLISHED_NETHER_CHITIN_STAIRS = AVPBlockItems.register(
         "polished_nether_chitin_stairs",
         AlienChitinBlocks.POLISHED_NETHER_CHITIN_STAIRS
+    );
+
+    public static final AVPDeferredHolder<BlockItem> POLISHED_NETHER_CHITIN_WALL = AVPBlockItems.register(
+        "polished_nether_chitin_wall",
+        AlienChitinBlocks.POLISHED_NETHER_CHITIN_WALL
     );
 
     public static void initialize() {}